move field factory class and minor updates
