+++ redisvl/schema/schema.py
@@ -2,57 +2,16 @@ import re
 import yaml
 from enum import Enum
 from pathlib import Path
-from typing import Any, Dict, List, Union, Tuple, Optional, Type
+from typing import Any, Dict, List
 
 from pydantic import BaseModel, ValidationError
 
-from redisvl.schema.fields import (
-    BaseField,
-    TagField,
-    TextField,
-    NumericField,
-    FlatVectorField,
-    HNSWVectorField,
-    GeoField
-)
+from redisvl.schema.fields import BaseField, FieldFactory
 
 
 class StorageType(Enum):
     HASH = "hash"
     JSON = "json"
-
-
-def get_vector_type(**field_data: Dict[str, Any]) -> Union[FlatVectorField, HNSWVectorField]:
-    """Get the vector field type from the field data."""
-
-    vector_field_classes = {
-        'flat': FlatVectorField,
-        'hnsw': HNSWVectorField
-    }
-    algorithm = field_data.get('algorithm', '').lower()
-    if algorithm not in vector_field_classes.keys():
-        raise ValueError(f"Unknown vector field algorithm: {algorithm}")
-
-    # default to FLAT
-    return vector_field_classes.get(algorithm, FlatVectorField)(**field_data)
-
-class FieldFactory:
-    FIELD_TYPE_MAP = {
-        "tag": TagField,
-        "text": TextField,
-        "numeric": NumericField,
-        "geo": GeoField,
-        "vector": get_vector_type
-    }
-
-    @staticmethod
-    def create_field(field_type: str, name: str, **kwargs) -> BaseField:
-        field_class = FieldFactory.FIELD_TYPE_MAP.get(field_type)
-        if not field_class:
-            raise ValueError(f"Unknown field type: {field_type}")
-        return field_class(name=name, **kwargs)
-
-
 
 
 class IndexSchema(BaseModel):
@@ -109,6 +68,7 @@ class IndexSchema(BaseModel):
         name = kwargs.get('name', None)
         if name is None:
             raise ValueError("Field name is required.")
+
         new_field = FieldFactory.create_field(field_type, **kwargs)
         if any(field.name == name for field in self.fields.get(field_type, [])):
             raise ValueError(
@@ -154,14 +114,16 @@ class IndexSchema(BaseModel):
         which makes it tedious to manually define each field. This method
         can be used to automatically generate fields from a sample of data.
 
-        Note: Vector fields are not generated by this method
-        Note: This method is a hueristic and may not always generate the
+        Note: Vector fields are not generated by this method.
+        Note: This method is a heuristic and may not always generate the
             correct field type.
 
         Args:
             data (Dict[str, Any]): The sample data to generate fields from.
-            strict (bool): Whether to raise an error if a field type cannot be inferred.
-            ignore_fields (List[str]): A list of field names to ignore.
+            strict (bool, optional): Whether to raise an error if a field type
+                cannot be inferred. Defaults to False.
+            ignore_fields (List[str], optional): A list of field names to
+                ignore. Defaults to [].
 
         Returns:
             Dict[str, List[Dict[str, Any]]]: A dictionary of fields.
